+++ packages/scaffold/src/client.ts
@@ -198,6 +198,10 @@ export class Web3ModalScaffold {
       AccountController.setAddressExplorerUrl(addressExplorerUrl)
     }
 
+  protected setSmartAccountDeployed: (typeof AccountController)['setSmartAccountDeployed'] =
+    isDeployed => {
+      AccountController.setSmartAccountDeployed(isDeployed)
+    }
   // -- Private ------------------------------------------------------------------
   private async initControllers(options: ScaffoldOptions) {
     NetworkController.setClient(options.networkControllerClient)