+++ apps/laboratory/CHANGELOG.md
@@ -1,5 +1,29 @@
 # @apps/laboratory
 
+## 4.0.12
+
+### Patch Changes
+
+- [#2014](https://github.com/WalletConnect/web3modal/pull/2014) [`95b35e1`](https://github.com/WalletConnect/web3modal/commit/95b35e1ebaf261a56a29cd9254d85b7c1430bfc0) Thanks [@tomiir](https://github.com/tomiir)! - Smart Account RPC handler canary
+
+- Smart Account initialization and feature flag
+
+- Updated dependencies [[`95b35e1`](https://github.com/WalletConnect/web3modal/commit/95b35e1ebaf261a56a29cd9254d85b7c1430bfc0)]:
+  - @web3modal/ethers@4.0.12
+  - @web3modal/wagmi@4.0.12
+  - @web3modal/siwe@4.0.12
+
+## 4.0.12-0c59f84f.0
+
+### Patch Changes
+
+- Smart Account RPC handler canary
+
+- Updated dependencies []:
+  - @web3modal/ethers@4.0.12-0c59f84f.0
+  - @web3modal/wagmi@4.0.12-0c59f84f.0
+  - @web3modal/siwe@4.0.12-0c59f84f.0
+
 ## 4.0.11
 
 ### Patch Changes